+++ src/main/java/dev/lambdaurora/spruceui/SpruceTextures.java
@@ -16,7 +16,7 @@ import net.minecraft.resources.Identifier;
  * Contains the identifiers of various useful textures.
  *
  * @author LambdAurora
- * @version 5.1.0
+ * @version 6.0.0
  * @since 5.1.0
  */
 public final class SpruceTextures {
@@ -32,18 +32,21 @@ public final class SpruceTextures {
 	/**
 	 * The dirt background texture used in pre-1.20.5 versions.
 	 */
-	public static final Identifier LEGACY_OPTIONS_BACKGROUND = Identifier.of("spruceui", "textures/gui/legacy_options_background.png");
+	public static final Identifier LEGACY_OPTIONS_BACKGROUND = SpruceUI.id("textures/gui/legacy_options_background.png");
 
 	/* Border */
 
+	public static final Identifier SIMPLE_BORDER_SPRITE = SpruceUI.id("border/simple");
+	public static final Identifier SIMPLE_HIGHLIGHTED_BORDER_SPRITE = SpruceUI.id("border/simple_highlighted");
+
 	public static final Identifier MENU_TOP_BORDER = Screen.HEADER_SEPARATOR;
 	public static final Identifier INWORLD_MENU_TOP_BORDER = Screen.INWORLD_HEADER_SEPARATOR;
-	public static final Identifier MENU_TOP_RIGHT_BORDER = Identifier.of("spruceui", "textures/gui/top_right_border_separator.png");
-	public static final Identifier INWORLD_MENU_TOP_RIGHT_BORDER = Identifier.of("spruceui", "textures/gui/inworld_top_right_border_separator.png");
-	public static final Identifier MENU_RIGHT_BORDER = Identifier.of("spruceui", "textures/gui/right_border_separator.png");
-	public static final Identifier INWORLD_MENU_RIGHT_BORDER = Identifier.of("spruceui", "textures/gui/inworld_right_border_separator.png");
-	public static final Identifier MENU_BOTTOM_RIGHT_BORDER = Identifier.of("spruceui", "textures/gui/bottom_right_border_separator.png");
-	public static final Identifier INWORLD_MENU_BOTTOM_RIGHT_BORDER = Identifier.of("spruceui", "textures/gui/inworld_bottom_right_border_separator.png");
+	public static final Identifier MENU_TOP_RIGHT_BORDER = SpruceUI.id("textures/gui/top_right_border_separator.png");
+	public static final Identifier INWORLD_MENU_TOP_RIGHT_BORDER = SpruceUI.id("textures/gui/inworld_top_right_border_separator.png");
+	public static final Identifier MENU_RIGHT_BORDER = SpruceUI.id("textures/gui/right_border_separator.png");
+	public static final Identifier INWORLD_MENU_RIGHT_BORDER = SpruceUI.id("textures/gui/inworld_right_border_separator.png");
+	public static final Identifier MENU_BOTTOM_RIGHT_BORDER = SpruceUI.id("textures/gui/bottom_right_border_separator.png");
+	public static final Identifier INWORLD_MENU_BOTTOM_RIGHT_BORDER = SpruceUI.id("textures/gui/inworld_bottom_right_border_separator.png");
 	public static final Identifier MENU_BOTTOM_BORDER = Screen.FOOTER_SEPARATOR;
 	public static final Identifier INWORLD_MENU_BOTTOM_BORDER = Screen.INWORLD_FOOTER_SEPARATOR;
 }