+++ src/main/java/dev/lambdaurora/spruceui/background/MenuBackground.java
@@ -9,12 +9,12 @@
 
 package dev.lambdaurora.spruceui.background;
 
-import com.mojang.blaze3d.systems.RenderSystem;
 import dev.lambdaurora.spruceui.SpruceTextures;
 import dev.lambdaurora.spruceui.widget.SpruceWidget;
 import dev.lambdaurora.spruceui.widget.WithBorder;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
+import net.minecraft.client.renderer.RenderType;
 import net.minecraft.resources.Identifier;
 
 /**
@@ -23,7 +23,7 @@ import net.minecraft.resources.Identifier;
  * @param texture the texture used for the background
  * @param inWorldTexture the textured used for the background when playing in a world
  * @author LambdAurora
- * @version 5.1.0
+ * @version 6.0.0
  * @since 5.1.0
  */
 public record MenuBackground(Identifier texture, Identifier inWorldTexture) implements Background {
@@ -38,8 +38,6 @@ public record MenuBackground(Identifier texture, Identifier inWorldTexture) impl
 	public void render(GuiGraphics graphics, SpruceWidget widget, int vOffset, int mouseX, int mouseY, float delta) {
 		int x = widget.getX();
 		int y = widget.getY();
-		int endX = widget.getEndX();
-		int endY = widget.getEndY();
 		int width = widget.getWidth();
 		int height = widget.getHeight();
 
@@ -48,22 +46,18 @@ public record MenuBackground(Identifier texture, Identifier inWorldTexture) impl
 
 			x += border.getLeft();
 			y += border.getTop();
-			endX -= border.getRight();
-			endY -= border.getBottom();
 
 			width -= border.getLeft() + border.getRight();
 			height -= border.getTop() + border.getBottom();
 		}
 
-		RenderSystem.enableBlend();
 		Identifier identifier = CLIENT.level == null ? this.inWorldTexture : this.texture;
 		graphics.drawTexture(
-				identifier,
+				RenderType::guiTextured, identifier,
 				x, y,
-				endX, endY,
+				0, 0,
 				width, height,
-				32, 32
+				16, 16
 		);
-		RenderSystem.disableBlend();
 	}
 }