+++ src/main/java/dev/lambdaurora/spruceui/border/SimpleBorder.java
@@ -9,28 +9,23 @@
 
 package dev.lambdaurora.spruceui.border;
 
-import com.mojang.blaze3d.systems.RenderSystem;
-import com.mojang.blaze3d.vertex.*;
 import dev.lambdaurora.spruceui.util.ColorUtil;
 import dev.lambdaurora.spruceui.widget.SpruceWidget;
 import net.minecraft.client.gui.GuiGraphics;
-import net.minecraft.client.renderer.GameRenderer;
-
-import java.util.Arrays;
 
 /**
  * Represents a simple solid border to draw around a widget.
  *
  * @author LambdAurora
- * @version 5.0.0
- * @since 2.0.0
+ * @version 6.0.0
+ * @since 6.0.0
  */
 public final class SimpleBorder implements Border {
 	public static final SimpleBorder SIMPLE_BORDER = new SimpleBorder(1, 192, 192, 192, 255);
 
 	private final int thickness;
-	private final int[] color;
-	private final int[] focusedColor;
+	private final int color;
+	private final int focusedColor;
 
 	public SimpleBorder(int thickness, int color) {
 		this(thickness, color, color);
@@ -38,8 +33,8 @@ public final class SimpleBorder implements Border {
 
 	public SimpleBorder(int thickness, int color, int focusedColor) {
 		this.thickness = thickness;
-		this.color = ColorUtil.unpackARGBColor(color);
-		this.focusedColor = ColorUtil.unpackARGBColor(focusedColor);
+		this.color = color;
+		this.focusedColor = color;
 	}
 
 	public SimpleBorder(int thickness, int red, int green, int blue, int alpha) {
@@ -48,50 +43,25 @@ public final class SimpleBorder implements Border {
 
 	public SimpleBorder(int thickness, int red, int green, int blue, int alpha, int focusedRed, int focusedGreen, int focusedBlue, int focusedAlpha) {
 		this.thickness = thickness;
-		this.color = new int[]{red, green, blue, alpha};
-		this.focusedColor = new int[]{focusedRed, focusedGreen, focusedBlue, focusedAlpha};
+		this.color = ColorUtil.packARGBColor(red, green, blue, alpha);
+		this.focusedColor = ColorUtil.packARGBColor(focusedRed, focusedGreen, focusedBlue, focusedAlpha);
 	}
 
 	@Override
 	public void render(GuiGraphics graphics, SpruceWidget widget, int mouseX, int mouseY, float delta) {
-		var tessellator = Tessellator.getInstance();
-		var buffer = tessellator.begin(VertexFormat.Mode.QUADS, DefaultVertexFormat.POSITION_COLOR);
-		RenderSystem.setShader(GameRenderer::getPositionColorShader);
 		int x = widget.getX();
 		int y = widget.getY();
 		int right = x + widget.getWidth();
 		int bottom = y + widget.getHeight();
-		boolean focused = widget.isFocused();
+		int color = widget.isFocused() ? this.focusedColor : this.color;
 		// Top border
-		this.vertex(buffer, x, y + this.thickness, focused);
-		this.vertex(buffer, right, y + this.thickness, focused);
-		this.vertex(buffer, right, y, focused);
-		this.vertex(buffer, x, y, focused);
+		graphics.fill(x, y, right, y + this.thickness, color);
 		// Right border
-		this.vertex(buffer, right - this.thickness, bottom, focused);
-		this.vertex(buffer, right, bottom, focused);
-		this.vertex(buffer, right, y, focused);
-		this.vertex(buffer, right - this.thickness, y, focused);
+		graphics.fill(right - this.thickness, y, right, bottom, color);
 		// Bottom
-		this.vertex(buffer, x, bottom, focused);
-		this.vertex(buffer, right, bottom, focused);
-		this.vertex(buffer, right, bottom - this.thickness, focused);
-		this.vertex(buffer, x, bottom - this.thickness, focused);
+		graphics.fill(x, bottom, right, bottom - this.thickness, color);
 		// Left border
-		this.vertex(buffer, x, bottom, focused);
-		this.vertex(buffer, x + this.thickness, bottom, focused);
-		this.vertex(buffer, x + this.thickness, y, focused);
-		this.vertex(buffer, x, y, focused);
-		MeshData builtBuffer = buffer.build();
-		if (builtBuffer != null) {
-			BufferUploader.drawWithShader(builtBuffer);
-		}
-		tessellator.clear();
-	}
-
-	private void vertex(BufferBuilder buffer, int x, int y, boolean focused) {
-		int[] color = focused ? this.focusedColor : this.color;
-		buffer.addVertex(x, y, 0).color(color[0], color[1], color[2], color[3]);
+		graphics.fill(x, y, x + this.thickness, bottom, color);
 	}
 
 	@Override
@@ -103,8 +73,8 @@ public final class SimpleBorder implements Border {
 	public String toString() {
 		return "SimpleBorder{" +
 				"thickness=" + this.thickness +
-				", color=" + Arrays.toString(this.color) +
-				", focusedColor=" + Arrays.toString(this.focusedColor) +
+				", color=" + Integer.toHexString(this.color) +
+				", focusedColor=" + Integer.toHexString(this.focusedColor) +
 				'}';
 	}
 }